api/requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
beautifulsoup4==4.10.0
bleach==4.1.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
decorator==5.1.1
dill==0.3.4
exif==1.3.5
expiringdict==1.2.1
fastapi==0.73.0
frozendict==2.3.0
google-api-core==2.5.0
google-auth==2.6.0
google-cloud==0.34.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
gunicorn==20.1.0
h11==0.13.0
html5lib==1.1
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
isodate==0.6.1
lxml==4.7.1
Markdown==3.3.6
# mdx_linkify==2.1
markdown-urlize==0.2.0
# markdown-customblocks==1.1.1
nose==1.3.7
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pygments==2.11.2
PyLD==2.0.3
PyJWT==1.7.1
pymdown-extensions==9.1
pyparsing==3.0.7
pytest==7.0.0
python-dateutil==2.8.2
PyYAML==6.0
rdflib==6.1.1
rdflib-jsonld==0.6.2
rednose==1.3.0
requests==2.27.1
responses==0.18.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
termstyle==0.1.11
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.4
webencodings==0.5.1
yamlns==0.9.2
zipp==3.7.0
git+https://github.com/rdsnyder/mdx_outline.git
git+https://github.com/rdsnyder/markdown-customblocks.git
